Add About Me wipe animation and structure.
The About Me room is composed of a header and the content. Both have
their own animations. The html and css of the header are fairly
straightforward, but those of the content can be a bit confusing. The
content has two parts to it, the text and the background. It is
necessary that these are two different elements because we want the
height and width to be "auto", but including this in an animation or
transition causes unpredictable effects. Therefore, I made the
background element a child of the text element and made the final values
of the height and width 100% with respect to the text element. This way,
we can have height and width animations that end in a size that is
dependent on the content.